테스트 코드 작성에 오류가 있었다. Tx을 걸고 작업하게 되면 Tx이 완료되어야만 캐시에도 작업이 수행된다. 이런 방식으로 테스트 코드를 변경해야 한다.

+++ hibernate-redis/src/main/java/org/hibernate/cache/redis/SingletonRedisRegionFactory.java
@@ -49,17 +49,13 @@ public class SingletonRedisRegionFactory extends AbstractRedisRegionFactory {
 
     @Override
     public void start(Settings settings, Properties properties) throws CacheException {
-        if (isDebugEnabled)
-            log.debug("Starting region factory...");
+        log.info("Redis를 2차 캐시 저장소로 사용하는 RedisRegionFactory를 시작합니다...");
 
         this.settings = settings;
-
         try {
             this.jedisClient = JedisTool.createJedisClient(props);
             ReferenceCount.incrementAndGet();
-
-            if (isDebugEnabled)
-                log.debug("Start region factory");
+            log.info("RedisRegionFactory를 시작했습니다!!!");
         } catch (Exception e) {
             throw new CacheException(e);
         }
@@ -67,18 +63,19 @@ public class SingletonRedisRegionFactory extends AbstractRedisRegionFactory {
 
     @Override
     public void stop() {
-        if (isDebugEnabled)
-            log.debug("Stop regoin factory...");
+        log.debug("RedisRegionFactory 사용을 중지합니다...");
 
         try {
-            if (ReferenceCount.decrementAndGet() == 0) {
-                jedisClient.flushDb();
+            if (jedisClient != null) {
+                if (ReferenceCount.decrementAndGet() == 0) {
+                    jedisClient.flushDb();
 
-                if (isDebugEnabled)
-                    log.debug("flush db");
+                    if (isDebugEnabled)
+                        log.debug("flush db");
+                }
+                log.info("RedisREgionFactory를 중지하였습니다.");
+                jedisClient = null;
             }
-            log.debug("Stop region factory is success");
-            jedisClient = null;
         } catch (Exception e) {
             log.error("jedisClient region factory fail to stop.", e);
         }
